refactor: domain 패키지 안에 vo 패키지 추가
